requirements.txt
PyYAML>=6.0
requests>=2.31
python-dotenv>=1.0
numpy>=1.24
pandas>=2.0

duckdb>=1.1
